requirements.txt
Flask==3.0.0
Flask-Babel==4.0.0
python-dotenv==1.0.0
colorlog==6.8.0
waitress==3.0.1
requests
Markdown
python-frontmatter
markdown-gfm-admonition
pymdown-extensions

